// src/api/types/schema/models/ApiCheckpointSessionDetailDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ApiCheckpointSessionDetailDto = {
    agent: string;
    context_text: string;
    created_at: string;
    is_task: boolean;
    metadata_json: string;
    prompts_text: string;
    session_id: string;
    session_index: number;
    tool_use_id: string;
    transcript_jsonl: string;
};